ch4/pyproject.toml
[project]
name = "rag-with-haystack-ch4"
version = "0.1"
description = "Building pipelines with Haystack"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
keywords = ["llm", "rag", "nlp", "pipelines"]
authors = [
  {name = "Laura Funderburk"}
]

classifiers = [
  "Development Status :: 1 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "haystack-ai>=2.19.0",
  "ipykernel>=7.0.1",
  "langchain-community>=0.4",
  "pandas>=2.3.3",
  "pymupdf>=1.26.5",
  "pypdf>=6.1.3",
  "python-dotenv>=1.1.1",
  "ragas>=0.3.7",
  "ragas-haystack>=1.0.0",
  "rapidfuzz>=3.14.1",
  "sentence-transformers>=5.0.0",
  "trafilatura>=2.0.0",
  "transformers[sentencepiece,torch]>=4.57.1",
]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {exclude = ["ch*", "venvs", "tests", "docs"]}}

[tool.pytest.ini_options]
addopts = "-v"
testpaths = [
  "pytests"
]

[tool.mypy]
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff.lint]
select = [
    "A",
    "B",
    "D",
    "DTZ",
    "E",
    "EM",
    "F",
    "I",
    "PL",
    "W",
]
ignore = [
    "D105",
    "E731",
    "PLR",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.per-file-ignores]
"examples/*.py" = ["D"]
"test_*.py" = ["D"]
